+++ src/main/java/com/mipt/hsse/hssetechbackend/payments/rent/PaymentCreateRentProcessor.java
@@ -3,6 +3,7 @@ package com.mipt.hsse.hssetechbackend.payments.rent;
 import com.mipt.hsse.hssetechbackend.auxiliary.VerificationResult;
 import com.mipt.hsse.hssetechbackend.data.entities.ClientTransactionStatus;
 import com.mipt.hsse.hssetechbackend.payments.RentCostCalculator;
+import com.mipt.hsse.hssetechbackend.payments.exceptions.WalletNotFoundException;
 import com.mipt.hsse.hssetechbackend.payments.exceptions.WalletUpdatingException;
 import com.mipt.hsse.hssetechbackend.payments.services.TransactionServiceBase;
 import com.mipt.hsse.hssetechbackend.payments.services.dto.TransactionInfo;
@@ -17,7 +18,7 @@ import java.util.Optional;
 @Component
 public class PaymentCreateRentProcessor implements CreateRentProcessor {
   private final TransactionServiceBase transactionService;
-
+  private static final String RENT_TRANSACTION_NAME = "Оплата аренды";
 
   public PaymentCreateRentProcessor(TransactionServiceBase transactionService) {
     this.transactionService = transactionService;
@@ -33,15 +34,17 @@ public class PaymentCreateRentProcessor implements CreateRentProcessor {
     var renter = rent.getRenter();
     var wallet = renter.getWallet();
 
-    var transName = "Аренда \"%s\"".formatted(targetItem.getDisplayName());
+    var transDescription = "Аренда \"%s\"".formatted(targetItem.getDisplayName()).intern();
 
     try {
-      var trans = transactionService.createTransaction(new TransactionInfo(rentCost, wallet.getId(), transName, Optional.empty()));
+      var trans = transactionService.createTransaction(new TransactionInfo(rentCost, wallet.getId(), RENT_TRANSACTION_NAME, Optional.of(transDescription)));
       transactionService.setTransactionStatus(trans.getId(), ClientTransactionStatus.SUCCESS);
 
       return VerificationResult.buildValid();
     } catch (WalletUpdatingException e) {
       return VerificationResult.buildInvalid("Not enough money");
+    } catch (WalletNotFoundException e) {
+      return VerificationResult.buildInvalid("Wallet not found");
     }
   }
 }
